requirements.txt
pychord==1.2.0
pyxdg==0.28
